Move the bcc source files into its own directory, 'bcc-deps'
This is to clarify the origin of these source files, and is meant to be
preparation for when we pull in more source files from BCC later on.

This also adds a README clarifying the source and purpose of these
files, stating the particular commit from the BCC repository which they
were pulled from.

Signed-off-by: Jazel Canseco <jcanseco@google.com>