Add caching mechanism to cache ProcSyms instances
This is to avoid the needless recreation of ProcSyms every time a
userspace symbol is looked up.

The key-value caching mechanism, which uses pid's as keys, also handles
the following pid collision cases: (1) a pid is reused by a new,
different process; and (2) the pid of a process goes over the default max
pid for Linux machines.

Signed-off-by: Jazel Canseco <jcanseco@google.com>