Modify CMakeLists to support dynamic linking for x86 builds
This adds support for bpfd builds that dynamically link to its
libraries. This is done only for x86 builds; arm64 builds continue to be
static as before.

This also consolidates the bcc functionality depended on by bpfd into one
library, 'libbpfd-bcc', as opposed to before where it was split into two
libraries, one for bpf-specific functionality ('libbpfd-bpf'), and one for
the rest of bcc depedencies ('libbpfd-bcc'). This was done since there
was little reason to keep them separate.

Signed-off-by: Jazel Canseco <jcanseco@google.com>