Remove kernel version flag handling
This removes the handling of the "-k" command line flag, which
corresponds to the "kernel version" of the device. This originally
existed since it was believed that BCC was passing the wrong kernel
version to BPFd via the BPF_PROG_LOAD command. It was believed that it
was passing the kernel version of the host instead of the target device.
However, testing has shown that BCC was actually passing the target
device's kernel version correctly. Therefore, there is no longer a
need for the "-k" flag.

Signed-off-by: Jazel Canseco <jcanseco@google.com>